[ci] Refactor code
- improves code docs
- move some generic utils to utils module
