+++ scripts/createSizeSnapshot.js
@@ -2,20 +2,10 @@
 const fse = require('fs-extra');
 const path = require('path');
 const getSize = require('size-limit');
+const { flatten, fromEntries } = require('./utils');
 
 const workspaceRoot = path.join(__dirname, '../');
 const snapshotDestPath = path.join(workspaceRoot, 'size-snapshot.json');
-
-function flatten(arr) {
-  return arr.reduce((acc, entry) => acc.concat(entry), []);
-}
-
-function fromEntries(entries) {
-  return entries.reduce((acc, [key, value]) => {
-    acc[key] = value;
-    return acc;
-  }, {});
-}
 
 /**
  * @param {object} snapshot snaphot generated by rollup-plugin-size-snapshot
@@ -35,6 +25,10 @@ async function getRollupSize(snapshotPath) {
   ]);
 }
 
+/**
+ * creates size snapshot for every bundle that is analyzed by `size-limit`
+ * @param {object} bundles
+ */
 function getSizeLimitSizes(bundles) {
   return bundles.map(async bundle => [
     bundle.name,
